problem was, that i dismissed to feed the learnrate in one sess.run, which is used while testtime..

+++ fingerYolo/training.py
@@ -584,16 +584,16 @@ def main():
             for i in range(nr_of_epochs/nr_of_epochs_until_save_model):
                 if i < 150000:
                     for j in range(nr_of_epochs_until_save_model):
-                        _ = sess.run([train_step],feed_dict={training: True, learnrate : np.float32(learning_rate)})
+                        _ = sess.run([train_step],feed_dict={training: True, learnrate : learning_rate})
                 elif i < 250000:
                     for j in range(nr_of_epochs_until_save_model):
-                        _ = sess.run([train_step],feed_dict={training: True, learnrate : (np.float32(learning_rate)/10)})  
+                        _ = sess.run([train_step],feed_dict={training: True, learnrate : (learning_rate/10)})  
                 elif i < 350000:
                     for j in range(nr_of_epochs_until_save_model):
-                        _ = sess.run([train_step],feed_dict={training: True, learnrate : (np.float32(learning_rate)/100)})  
+                        _ = sess.run([train_step],feed_dict={training: True, learnrate : (learning_rate/100)})  
                 else:
                     for j in range(nr_of_epochs_until_save_model):
-                        _ = sess.run([train_step],feed_dict={training: True, learnrate : (np.float32(learning_rate)/1000)})  
+                        _ = sess.run([train_step],feed_dict={training: True, learnrate : (learning_rate/1000)})  
                 
                 #testing in traindata while training
                 numbers_of_iterations_until_now = i*nr_of_epochs_until_save_model+j+1
